refactored buffer.rs

+++ src/buffer.rs
@@ -28,21 +28,21 @@ impl Buffer {
         self.lines.len()
     }
 
-    pub fn insert(&mut self, x: u16, y: u16, c: char) {
+    pub fn insert(&mut self, x: u16, y: usize, c: char) {
         if let Some(line) = self.lines.get_mut(y as usize) {
             (*line).insert(x as usize, c);
         }
     }
 
-    pub fn remove(&mut self, x: u16, y: u16) {
+    pub fn remove(&mut self, x: u16, y: usize) {
         if let Some(line) = self.lines.get_mut(y as usize) {
             (*line).remove(x as usize);
         }
     }
 
-    pub fn remove_line(&mut self, line: u16) {
-        if self.len() > line as usize {
-            self.lines.remove(line as usize);
+    pub fn remove_line(&mut self, line: usize) {
+        if self.len() > line {
+            self.lines.remove(line);
         }
     }
 }
